Added the functionality of saving as a single file or in multiple files.
Multiple file mode: This wil create and/or save to the folder, where each debtor will be saved and named after each debtor. Each file needs to start with "Debtor - " which will be used as an indicator for that it is a part of this application.

Single file mode: This will save the whole list of debtors including their information in one file. The inputbox should contain the directory AND filename.

Fixed a little bug/problem when saving file(s), and changed the icon to a bag with a dollar sign, instead of a demon with a halo on top.